Ble mer ryddig/mindre kodekopi å samle tester for InterntVedtakV2 i InterntVedtakServiceTest i stedet for en helt egen testklasse
